Add an FPS indicator and fix opening PNGs
There was no quick way to tell how the game was performing, so I added a function
to display the average FPS in the top-left corner of the screen. Also, the path
for opening some PNGs was wrong, so I changed it; "~/imgs/Interface/SomeImage.png"
- the word "Interface" is capitalized, but the actual directory is "interface". Hope
that "interface" was actually the intended one for use.