Allow game to progress through different levels
Keep track of the current_level and increment it after a victory.
Load the new map and allow the player to continue playing. Display
the current level in the sidebar to keep track. Had to modify the
return value of level.update_state() to allow for more than just
a binary state of quit or keep playing. Tried instead to use the
existing states of G, I and F, then added a state V for victory.