text - New function to display shopkeeper balance, modified STR_THIS_ITEM_CANT_BE_SOLD to better fit context shop - Added new variable shop_balance; initialized to 500 for testing, buy and sell functions adjusted to update vendor amount menu_creator_manager - create_shop_menu updated to show vendor gold in UI load xml, load tmx, level_scene - added shop_balance parameter to constructors
Additional Notes: Shop balance does not show on sell interface. Shop balance value is not updated when saved, may be due to the fact that they were set to 500 in the constructor. The load xml, load tmx, and level_scene files need to be reviewed for confirmation. The basic functionality is there, but tweaks still need to be made.